refactoring structs, enums and variable names
The library follows multiple naming schemes, which are refactored to the ESP-IDF
standard by this commit. On one hand, this will improve code readability, but on
the other hand, backporting bugfixes from the original source repository will be
more difficult. But this might not be a big problem as the original source code
got not touched for years.